Add transaction methods
