+++ crates/sdk/src/api/method/mod.rs
@@ -46,6 +46,7 @@ mod select;
 mod set;
 mod signin;
 mod signup;
+mod transaction;
 mod unset;
 mod update;
 mod upsert;
@@ -58,7 +59,6 @@ mod tests;
 
 pub use authenticate::Authenticate;
 pub use begin::Begin;
-pub use begin::Transaction;
 pub use cancel::Cancel;
 pub use commit::Commit;
 pub use content::Content;
@@ -83,6 +83,7 @@ pub use set::Set;
 pub use signin::Signin;
 pub use signup::Signup;
 use tokio::sync::watch;
+pub use transaction::Transaction;
 pub use unset::Unset;
 pub use update::Update;
 pub use upsert::Upsert;
